chore: consolidate duration parsing, submit flow, and k8s utils reuse
