+++ pkg/cli/run.go
@@ -6,7 +6,6 @@ import (
 
 	"github.com/spf13/cobra"
 
-	"github.com/kubeflow/arena/pkg/client"
 	"github.com/kubeflow/arena/pkg/constants"
 	"github.com/kubeflow/arena/pkg/log"
 	"github.com/kubeflow/arena/pkg/provider"
@@ -52,87 +51,7 @@ Use --set to override YAML fields with Helm-style dot-notation paths.`,
 
 		log.Debug("task loaded", "name", t.Name, "framework", t.Framework.Name)
 
-		// Get provider
-		p, err := getProvider(t.Framework.Name)
-		if err != nil {
-			log.Error(err, "failed to get provider", "framework", t.Framework.Name)
-			return err
-		}
-
-		// Create K8s client early for version detection
-		k8sClient, err := client.NewClient(kubeconfig, kubeContext)
-		if err != nil {
-			log.Error(err, "failed to create Kubernetes client")
-			return fmt.Errorf("failed to create Kubernetes client: %w", err)
-		}
-
-		// For MPI-family frameworks, detect cluster version
-		if isMPIFamily(t.Framework.Name) {
-			if mpiP, ok := p.(*provider.MPIProvider); ok {
-				version, err := resolveMPIAPIVersion(cmdContext(cmd), k8sClient)
-				if err != nil {
-					return err
-				}
-				mpiP.APIVersion = version
-			}
-		}
-
-		log.Debug("building CRD", "framework", t.Framework.Name)
-
-		// Build CRD
-		crd, err := p.BuildCRD(t)
-		if err != nil {
-			log.Error(err, "failed to build CRD", "framework", t.Framework.Name)
-			return fmt.Errorf("failed to build CRD: %w", err)
-		}
-
-		// Set namespace from resolved value
-		ns := resolveNS(t.Namespace)
-		crd.SetNamespace(ns)
-		t.Namespace = ns
-
-		setFrameworkLabel(crd, t.Framework.Name)
-
-		log.Debug("CRD built", "kind", crd.GetKind(), "name", crd.GetName(), "namespace", ns)
-
-		// Dry-run: print and exit
-		if runDryRun {
-			log.Info("dry-run mode, not submitting")
-			return printDryRun(crd, t)
-		}
-
-		log.Debug("checking if job exists", "name", t.Name, "namespace", ns)
-
-		// Check if job already exists
-		if err := checkJobExists(cmdContext(cmd), k8sClient, ns, t.Name); err != nil {
-			return err
-		}
-
-		log.Debug("submitting job", "kind", crd.GetKind(), "name", crd.GetName(), "namespace", ns)
-
-		// Submit to cluster
-		if err := k8sClient.Create(cmdContext(cmd), crd); err != nil {
-			log.Error(err, "job creation failed", "kind", crd.GetKind(), "name", crd.GetName())
-			return fmt.Errorf("failed to submit job: %w", err)
-		}
-
-		log.Debug("creating auxiliary resources", "name", crd.GetName())
-
-		// Create auxiliary resources (ConfigMap anchor, TensorBoard)
-		if err := createJobResources(cmdContext(cmd), crd, t, k8sClient, p); err != nil {
-			// Best-effort cleanup: delete the CRD so we don't leave an orphaned job
-			log.Warning("auxiliary resource creation failed, cleaning up CRD",
-				"kind", crd.GetKind(), "name", crd.GetName(), "error", err.Error())
-			if delErr := k8sClient.Delete(cmdContext(cmd), crd.GetKind(), ns, t.Name); delErr != nil {
-				log.Warning("failed to clean up CRD after partial failure",
-					"kind", crd.GetKind(), "name", crd.GetName(), "error", delErr.Error())
-			}
-			return err
-		}
-
-		log.Info("job submitted successfully", "name", t.Name, "namespace", ns)
-		fmt.Printf("Job %s submitted successfully\n", t.Name)
-		return nil
+		return submitCRD(cmdContext(cmd), t, t.Framework.Name, runDryRun)
 	},
 }
